split direct file read host effects
