+++ docs/COMPLETION_AUDIT.md
@@ -303,6 +303,11 @@ and do not assume Phase 4 is ready without evidence.
   `tests/integration/cli_build/legacy_graph_command_host_effects/file_reads.rs`,
   leaving env-effect ordering and unrelated test skip coverage in the parent
   module.
+- Direct `zen build.zen` host-effect integration tests now split declared
+  file-read fallback accept/reject coverage into
+  `tests/integration/cli_build/direct_build_graph_host_effects/file_reads.rs`,
+  leaving env-effect ordering and unrelated test skip coverage in the parent
+  module.
 - Build graph unit tests now split target metadata validation and gated
   package/link target rejection coverage into
   `tests/build_graph/target_metadata.rs`, leaving the root build graph test
